Generalize PlaceConverter
We will need the same logic for other components as we slowly migrate
to StableMIR. Move that to `ty_stable.rs` for now.
